The BusinessLayer - player-finding for starting a game of 4 [main idea]
thinking about a way to connect 2-4 Users currently searching for a game, and creating a new game containing 2-4 Player, one Player corresponding to each User.
The GameStartProcedure class will have a static User list or array, and every user sending a request to enter a game will be added to this static list/array. every time the list/array reaches the count of players wanted(2-4), it will create the corresponding amount of Players, Inheriting from those Users, and send back a game containing these 2-4 players in it, and reset the User List/array.

After that, each Player will be given a shuffled hand, ad the game will start.

[edited on Mac]